namespace tracer, rename files, global tracer singleton, extract tracer configuration
