+++ lib/lightstep/tracer/transports/transport_nil.rb
@@ -4,9 +4,6 @@ class TransportNil
   def initialize
   end
 
-  def ensure_connection(options)
-  end
-
   def flush_report(_auth, _report)
     nil
   end
