+++ lib/lightstep/tracer/version.rb
@@ -1,5 +1,5 @@
 module LightStep
-  module Tracer
+  class Tracer
     VERSION = '0.8.14'.freeze
   end
 end
